feat: improve <AfterDraf>
Make <AfterDraf> work on server. As well as allow users to create their
custom <AfterDraf>s, which wait for DRAF only a given number of times,
for example, to prevent FOUC, your component needs to wait for DRAF only
for the first time but subsequent re-renders can render children
synchronously.